pyproject.toml
[project]
name = "git-changes-view"
version = "0.1.0"
description = "Tree view of git changes with line counts"
requires-python = ">=3.11"
dependencies = [
    "click>=8.0",
    "gitpython>=3.1",
]

[project.scripts]
git-changes-view = "git_changes_view.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/git_changes_view"]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0",
    "ruff>=0.4",
]
